+++ src/album/album.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateAlbumDto } from 'src/album/album.dto';
-import { prismaErrorHandling } from 'src/utils/prisma-error-handling';
+import { prismaErrorHandling } from 'src/utils/prisma-errors';
 
 @Injectable()
 export class AlbumService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService) { }
 
   async findAll() {
     try {
