+++ src/artist/artist.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateArtistDto } from './artist.dto';
-import { prismaErrorHandling } from 'src/utils/prisma-error-handling';
+import { prismaErrorHandling } from 'src/utils/prisma-errors';
 
 @Injectable()
 export class ArtistService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService) { }
 
   async findAll() {
     try {
