+++ src/logger/logger-http.middleware.ts
@@ -1,6 +1,6 @@
 import { Injectable, Logger, NestMiddleware } from '@nestjs/common';
 import { Request, Response, NextFunction } from 'express';
-import { writeLogs } from 'src/utils/log-levels';
+import { writeLogs } from 'src/utils/logs';
 
 @Injectable()
 export default class LoggingService implements NestMiddleware {
@@ -12,6 +12,7 @@ export default class LoggingService implements NestMiddleware {
       const { statusCode, statusMessage } = response;
       const message = `${method} ${url} ${params.id ?? 'no params'} ${statusMessage}`;
       let type;
+
       if (statusCode >= 500) {
         type = 'error';
         writeLogs(`HTTP ${type} ${message}`);
