+++ src/main.ts
@@ -4,9 +4,9 @@ import { ValidationPipe } from '@nestjs/common';
 import { SwaggerModule } from '@nestjs/swagger';
 import { document } from './app/parse-yaml';
 import { PrismaClientExceptionFilter } from 'nestjs-prisma';
-import { MAP_ERRORS } from './utils/const';
-import { getLogLevels } from './utils/log-levels';
-import { processStdinHandler, uncaughtErrorHandler, unhandledRejectionHandler } from './utils/utils';
+import { PRISMA_ERR_TO_HTTP_CODES } from './utils/const';
+import { getLogLevels } from './utils/logs';
+import { processStdinHandler, uncaughtErrorHandler, unhandledRejectionHandler } from './utils/handling-process';
 import { AllExceptionsFilter } from './exceptions/all-exceptions';
 
 const { PORT, NODE_ENV } = process.env;
@@ -16,17 +16,16 @@ async function bootstrap() {
   process.on('unhandledRejection', unhandledRejectionHandler);
   process.stdin.on('data', processStdinHandler);
 
-  const isProduction = NODE_ENV === 'production';
-  const app = await NestFactory.create(AppModule, { logger: getLogLevels(isProduction) });
+  const app = await NestFactory.create(AppModule, { logger: getLogLevels(NODE_ENV === 'production') });
 
   const httpAdapterHost = app.get(HttpAdapterHost);
-  app.useGlobalFilters(new PrismaClientExceptionFilter(httpAdapterHost.httpAdapter, MAP_ERRORS));
+  app.useGlobalFilters(new PrismaClientExceptionFilter(httpAdapterHost.httpAdapter, PRISMA_ERR_TO_HTTP_CODES));
   app.useGlobalFilters(new AllExceptionsFilter(httpAdapterHost));
   app.useGlobalPipes(new ValidationPipe());
   app.enableShutdownHooks();
 
   SwaggerModule.setup('doc', app, document);
-  process.on('SIGINT', async () => await app.close());
+
   console.log(`App is listening on ${PORT}`);
 
   await app.listen(PORT);
