+++ src/utils/logs.ts
@@ -2,12 +2,8 @@ import { InternalServerErrorException } from '@nestjs/common';
 import { LogLevel } from '@nestjs/common/services/logger.service';
 import { appendFile } from 'node:fs/promises';
 
-export function getLogLevels(isProduction: boolean): LogLevel[] {
-  if (isProduction) {
-    return ['log', 'warn', 'error'];
-  }
-  return ['error', 'warn', 'log', 'verbose', 'debug'];
-}
+export const getLogLevels = (isProduction: boolean): LogLevel[] =>
+  isProduction ? ['log', 'warn', 'error'] : ['error', 'warn', 'log', 'verbose', 'debug'];
 
 export async function writeLogs(log: string) {
   try {
