+++ src/utils/prisma-errors.ts
@@ -8,7 +8,6 @@ import { PrismaException } from 'src/exceptions/prisma-exceptions';
 
 export function prismaErrorHandling(error: unknown) {
   if (error instanceof PrismaClientKnownRequestError) {
-    console.log(error.code);
     throw new PrismaException(error.message, error.code);
   }
   if (error instanceof PrismaClientUnknownRequestError || error instanceof PrismaClientValidationError) {
